requirements.txt
Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.6.0
